requirements.txt
langchain
langgraph
chromadb
sentence-transformers
openai
streamlit
fastapi
uvicorn
python-dotenv
langchain-groq
langchain_community
langchain-huggingface
langchain-chroma
langchain-google-genai
fastapi
uvicorn
streamlit
requests